# opt: add two set version of `(*EquivGroups).AreAllColsEquiv`