A 2-set version of the `AreAllColsEquiv` method of `EquivGroups` has
been added so that two sets can be checked for equality without having
to union the sets together and possible allocate new memory.

Epic: None

Release note: None