@mgartner 2025-01-06
I bisected a regression in `BenchmarkPhases/many.*/.*/[OptBuild.*|AssignPlaceholders.*]` to https://github.com/cockroachdb/cockroach/commit/15b7d7e2963214393acee02befcdfe106f29ad79. I profiled the difference, and it looked due to allocations here: https://github.com/cockroachdb/cockroach/blob/15b7d7e2963214393acee02befcdfe106f29ad79/pkg/sql/opt/props/func_dep.go#L1827

This addresses the ~3.4% regression in allocations, but not the 20% regression in latency.

```
name                                                        old time/op    new time/op    delta
Phases/many-columns-and-indexes-a/Simple/OptBuildNoNorm-16     207µs ± 1%     207µs ± 1%    ~     (p=0.780 n=10+9)
Phases/many-columns-and-indexes-a/Simple/OptBuildNorm-16       207µs ± 0%     207µs ± 1%    ~     (p=0.400 n=9+10)
Phases/many-columns-and-indexes-b/Simple/OptBuildNoNorm-16     221µs ± 1%     221µs ± 1%    ~     (p=0.604 n=9+10)
Phases/many-columns-and-indexes-b/Simple/OptBuildNorm-16       227µs ± 1%     228µs ± 1%    ~     (p=0.182 n=9+10)
Phases/many-columns-and-indexes-c/Simple/OptBuildNorm-16      5.11ms ± 0%    5.11ms ± 1%    ~     (p=0.912 n=10+10)
Phases/many-columns-and-indexes-d/Simple/OptBuildNoNorm-16    5.40ms ± 0%    5.40ms ± 1%    ~     (p=0.720 n=9+10)
Phases/many-columns-and-indexes-d/Simple/OptBuildNorm-16      5.45ms ± 0%    5.46ms ± 1%    ~     (p=0.853 n=10+10)
Phases/batch-insert-many/Simple/OptBuildNoNorm-16             2.25ms ± 1%    2.26ms ± 1%    ~     (p=0.105 n=10+10)
Phases/many-columns-and-indexes-c/Simple/OptBuildNoNorm-16    4.85ms ± 0%    4.87ms ± 1%  +0.46%  (p=0.002 n=10+10)
Phases/batch-insert-many/Simple/OptBuildNorm-16               3.43ms ± 1%    3.45ms ± 0%  +0.57%  (p=0.001 n=9+9)

name                                                        old alloc/op   new alloc/op   delta
Phases/many-columns-and-indexes-c/Simple/OptBuildNoNorm-16    2.88MB ± 0%    2.81MB ± 0%  -2.23%  (p=0.000 n=10+10)
Phases/many-columns-and-indexes-d/Simple/OptBuildNoNorm-16    2.92MB ± 0%    2.85MB ± 0%  -2.20%  (p=0.000 n=10+10)
Phases/many-columns-and-indexes-d/Simple/OptBuildNorm-16      2.98MB ± 0%    2.91MB ± 0%  -2.16%  (p=0.000 n=8+10)
Phases/many-columns-and-indexes-c/Simple/OptBuildNorm-16      3.04MB ± 0%    2.97MB ± 0%  -2.12%  (p=0.000 n=10+10)
Phases/many-columns-and-indexes-a/Simple/OptBuildNoNorm-16    40.8kB ± 0%    40.8kB ± 0%    ~     (p=0.879 n=10+10)
Phases/many-columns-and-indexes-a/Simple/OptBuildNorm-16      43.2kB ± 0%    43.2kB ± 0%    ~     (p=0.775 n=10+9)
Phases/many-columns-and-indexes-b/Simple/OptBuildNoNorm-16    45.9kB ± 0%    45.9kB ± 0%    ~     (p=0.691 n=10+10)
Phases/many-columns-and-indexes-b/Simple/OptBuildNorm-16      50.2kB ± 0%    50.2kB ± 0%    ~     (p=0.587 n=10+10)
Phases/batch-insert-many/Simple/OptBuildNoNorm-16             3.36MB ± 0%    3.36MB ± 0%    ~     (p=0.725 n=10+10)
Phases/batch-insert-many/Simple/OptBuildNorm-16               3.99MB ± 0%    3.99MB ± 0%    ~     (p=0.447 n=10+10)

name                                                        old allocs/op  new allocs/op  delta
Phases/many-columns-and-indexes-c/Simple/OptBuildNoNorm-16     60.0k ± 0%     58.0k ± 0%  -3.35%  (p=0.000 n=10+10)
Phases/many-columns-and-indexes-d/Simple/OptBuildNoNorm-16     60.8k ± 0%     58.8k ± 0%  -3.31%  (p=0.000 n=10+10)
Phases/many-columns-and-indexes-d/Simple/OptBuildNorm-16       61.3k ± 0%     59.3k ± 0%  -3.28%  (p=0.000 n=9+9)
Phases/many-columns-and-indexes-c/Simple/OptBuildNorm-16       62.4k ± 0%     60.4k ± 0%  -3.22%  (p=0.000 n=10+10)
Phases/many-columns-and-indexes-a/Simple/OptBuildNoNorm-16      67.0 ± 0%      67.0 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-a/Simple/OptBuildNorm-16        74.0 ± 0%      74.0 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-b/Simple/OptBuildNoNorm-16       106 ± 0%       106 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-b/Simple/OptBuildNorm-16         128 ± 0%       128 ± 0%    ~     (all equal)
Phases/batch-insert-many/Simple/OptBuildNoNorm-16              18.7k ± 0%     18.7k ± 0%    ~     (all equal)
Phases/batch-insert-many/Simple/OptBuildNorm-16                29.4k ± 0%     29.4k ± 0%    ~     (p=0.231 n=10+6)
```